# Faster iterative linesearch