Still a niggling numerical instability but this is mostly working, will fix broken tests tomorrow.

I suspect the things that contributed the most to the speedup are:

- fusing some kernel launches
- using fewer atomic adds
- implemented very simple early exiting

Results of `mjx-testspeed --function=solve --is_sparse=False --mjcf=humanoid/humanoid.xml --batch_size=8192`

This PR:

```
 Total JIT time: 0.23 s
 Total simulation time: 1.07 s
 Total steps per second: 7,659,559
 Total realtime factor: 38,297.79 x
 Total time per step: 130.56 ns
```

Main:

```
Summary for 8192 parallel rollouts

 Total JIT time: 0.52 s
 Total simulation time: 1.27 s
 Total steps per second: 6,431,488
 Total realtime factor: 32,157.44 x
 Total time per step: 155.49 ns
```

Early exiting seems to be working quite nicely!   🎉    There is only a modest slowdown for 50 ls iterations with early exiting, compared to not early exiting.

Results of `mjx-testspeed --function=solve --is_sparse=False --mjcf=humanoid/humanoid.xml --batch_size=8192 --ls_iterations=20`

This PR:

```
 Total JIT time: 0.27 s
 Total simulation time: 1.14 s
 Total steps per second: 7,215,686
 Total realtime factor: 36,078.43 x
 Total time per step: 138.59 ns
```

Main:

```
 Total JIT time: 3.94 s
 Total simulation time: 4.30 s
 Total steps per second: 1,904,803
 Total realtime factor: 9,524.02 x
 Total time per step: 524.99 ns
```